1.2_feat_covar-identity-distance
22.9.2020. Update `transect_clustering` to calculate the correlation between
pairwise identity and distance.

I am not happy with the current measure of spatial clustering
(dividing mean pairwise distance between non-identical pairs by mean pairwise
distance between identical pairs) because it does not account for the absolute
number of identical pairs, so it blows up if there is a small number of identical
pairs.

Other measures I have found seem to be designed for continuous variables and do
but treat distance as discrete (nearest neighbour or not) (for example, Moran's I,
spatial entropy). In this case the outcome is discrete (identical or not) but
the distance is continuous.

The best idea I can come up with is to model spatial
clustering as the (Pearson; using ranks gives ties) correlation between pairwise
identity and distance.
This is not quite as neat as proper spatial autocorrelation stats, because these
return 1 for perfect clustering and -1 for perfect dispersion. The correlation
coeficient becomes negative for clustering but doesn't get much above zero for
perfect dispersion.

 On branch 1.2_feat_covar-identity-distance
 Changes to be committed:
	modified:   DESCRIPTION
	modified:   R/transect_clustering.R