23.9.2020. Replace population_size argument with how much bigger the population should be than the transect.
Currently there are arguments for population size and density, and area is
determined to be size/density. However, since transect length is fixed it can
happen that if population size is small, the width of the torus is less than
the length of the transect. Even if that is not the case, it means that the
distance from edge of the torus to the ends of the transect varies with
population size and density.

To fix this, remove the argument for population size and instead say that the
width of the torus will be 1.5 times the length of the transect, and define
population size as density * width^2. Return population size in the output. This
should affect `sim_population()` and `simmiad()`.

 Changes to be committed:
	modified:   DESCRIPTION
	modified:   R/sim_population.R
	new file:   history
	modified:   man/sim_population.Rd
	modified:   man/simmiad.Rd
	modified:   man/transect_clustering.Rd
	new file:   tests/testthat/test_sim_population.R