23.9.2020 Added `distance_identities` to calculates the probability of observing identical genotypes at different distances along a transect.
This originally aimed to address how what we can say about transect A-West, but
is informative more generally. For observed and simulated transects calculate
the probability of observing identical genotypes at sampling points at different
distance classes. You can then plot this as a decaying curve so you can make
statements like "we don't see identical genotypes more than x metres apart".
This can be compared with simulations, which I expect will be much flatter and
lower. This might prove to be a more robust measure of spatial differentiation
than a single number for spatial clustering, but I will calculate both for now.

You can also calculate the probability of observing identical genotypes in
corresponding sampling points 20m away in transect A-West, which I expect will
be higher than you'd get for 20m along either transect or at 20m in simulated
transects.

There is an obvious caveat that estimates for lower distance classes will be
better because there are more comparisons that can be made.

For observed transects, you probably need to average over all sampling years,
because at higher distances things will collapse into a small number of zeroes
and ones.
For simulated transects, estimate this for each year, and return a matrix with
a row for each replicate.
For the A/A-West comparison calculate curves for each transect in 20, plus a
mean for the average between-transect comparison at 20m.

 On branch 1.6_feat_identity_at_different_distances
 Changes to be committed:
	modified:   DESCRIPTION
	new file:   R/distance_identities.R
	new file:   tests/testthat/test_distance_identities.R