Added 'mixing' parameter v1.11.7
When simulating sturctured populations with initialise_population I need to be
able to change the scale of structure. If method='clusters' this is set via
mean_dispersal_distance. If method='mvnorm' this was hard coded to
`sqrt(habitat_size/n_starting_genotypes) * 3`.

For both instances of `method` I added parameter `mixing` to control this
behaviour. This replaces `mean_dispersal_distance` and the 3 in the snippet
above.

Updated sim_populatrion and simmiad accordingly.

 On branch master
 Changes to be committed:
	modified:   DESCRIPTION
	modified:   R/initialise_population.R
	modified:   R/sim_population.R
	modified:   R/simmiad.R
	modified:   history
	modified:   tests/testthat/test_initialise_population.R
	modified:   tests/testthat/test_simmiad.R
	modified:   tests/testthat/test_update_population.R