Fixed docs; mvtnorm to Suggests v1.12.1
I (hopefully) fixed a bug in the doc strings that meant the package couldn't
build.

Changed the dependency on mvtnorm from 'Depends' to 'Suggests', since this was
not playing nicely with different versions of R, and I am unlikely to use it
anyway.

Consider removing the extra methods for simulating structured populations if
using a real vector works.

 On branch master
 Changes to be committed:
	modified:   DESCRIPTION
	modified:   R/initialise_population.R
	modified:   R/sim_population.R
	modified:   history
	modified:   man/initialise_population.Rd
	modified:   man/parameter_table.Rd
	modified:   man/sim_population.Rd
	modified:   man/simmiad.Rd