Added di_by_year
v1.13.0

To look at the sturctured simulations I need to be able to track the changes
in clustering through time. I run distance_identities on the population for each
year separately, and save the value for the closest distance class only. This
returns a matrix with a row for each simulation replicate and a column for each
generation.

 On branch master
 Changes to be committed:
	modified:   DESCRIPTION
	modified:   R/simmiad.R
	modified:   history
	modified:   tests/testthat/test_simmiad.R
	modified:   tests/testthat/test_write_simmiad.R