Fixed which years to sample v1.14.1
Updated simmiad.R to require an integer list of which generations should be
'sampled' to reflect real data collection (for example only the last 36
generations). Spatial clustering is calculated over those years, and temporal
stability is calculated for each year by comparing to the first year in the
range. This is controlled by the argument `years_to_study` which replaces
arguments `stability_years` and `clustering_years`. `di_by_year` is still
calculated for each year in the series.

Updated parameter_table to reflect this.

 On branch master
 Changes to be committed:
	modified:   R/parameter_table.R
	modified:   R/simmiad.R
	modified:   history
	modified:   tests/testthat/test_simmiad.R
	modified:   DESCRIPTION