1.16.3 Variance in fitness
5th June 2023

The implementation in v1.16.0 isn't appropriate because the selection gradient
is on the logit scale, with the logistic transformation happening *before*
dividing by mean fitness. In this way, this really reflectin selection correctly.

Following Morrissey & Bonnet (2019; J. Heredity 110:396–402) I changed this so
that initialise_population generates log fitness values with mean zero and
variance var_w. This is exponentiated to give absolute fitness, before correcting
by mean fitness. In this way the strength of selection is controlled by var_w.

 Changes to be committed:
	modified:   DESCRIPTION
	modified:   NAMESPACE
	modified:   R/initialise_population.R
	modified:   R/parameter_table.R
	modified:   R/relative_fitness.R
	modified:   R/sim_population.R
	modified:   R/simmiad.R
	modified:   R/update_population.R
	modified:   history
	new file:   man/habitat_clustering.Rd
	modified:   man/initialise_population.Rd
	modified:   man/parameter_table.Rd
	new file:   man/relative_fitness.Rd
	modified:   man/sim_population.Rd
	modified:   man/simmiad.Rd
	modified:   man/update_population.Rd
	new file:   tests/testthat/test_relative_fitness.R
	modified:   tests/testthat/test_update_population.R
	modified:   tests/testthat/test_write_simmiad.R