1.16.4 Update phenotypes for outcrossed plants
6th June 2023
in update_population a subset of plants are randomly assigned to be outcrossed
and receive a new genotype. I updated the function to also assign a new
phenotype value. This is drawn from a normal distribution with a mean equal to
the mean of the current population and a sd equal to half the sd of the current
population (under the infinitesimal half the variance segregates between familes
and half within).

 On branch master
 Changes to be committed:
	modified:   DESCRIPTION
	modified:   R/update_population.R
	modified:   history